fetches user balance when they open the menu
