+++ src/hooks/useErc20.ts
@@ -3,7 +3,7 @@ import { erc20 } from '@/config/abis'
 import { formatContractResult } from '@/utils/LockOverviewUtils'
 import { ContractInterface } from '@ethersproject/contracts'
 import { BigNumber } from 'ethers'
-import { useAccount, useContractRead } from 'wagmi'
+import { useAccount, useBalance, useContractRead } from 'wagmi'
 
 const readContract = {
   addressOrName: config.iqAddress,
@@ -13,12 +13,9 @@ const readContract = {
 export const useErc20 = () => {
   const { address } = useAccount()
 
-  const { data: erc20Balance } = useContractRead({
-    ...readContract,
-    functionName: 'balanceOf',
-    args: [address],
-    watch: true,
-    staleTime: 5000,
+  const { data: erc20Balance } = useBalance({
+    addressOrName: address,
+    token: config.iqAddress,
   })
 
   const { data: totalValueLocked } = useContractRead({
@@ -28,7 +25,7 @@ export const useErc20 = () => {
   })
 
   const getUserBalance = () => {
-    return (erc20Balance as unknown as BigNumber) ?? 0
+    return (erc20Balance?.value as unknown as BigNumber) ?? 0
   }
 
   const tvl = () => {
