+++ src/hooks/useLockOverview.ts
@@ -36,8 +36,8 @@ export const useLockOverview = () => {
     functionName: 'locked__end',
     args: [address],
     overrides: { gasLimit: DEFAULT_GAS_LIMIT },
-    watch: true,
-    staleTime: 5000,
+    // watch: true,
+    // staleTime: 5000,
   })
 
   const {
@@ -49,8 +49,8 @@ export const useLockOverview = () => {
     functionName: 'locked',
     args: [address],
     overrides: { gasLimit: DEFAULT_GAS_LIMIT },
-    watch: true,
-    staleTime: 5000,
+    // watch: true,
+    // staleTime: 5000,
   })
 
   const getTotalHiiqSupply = () => {
